feat(api): add another way of customizing otp query
This option allows setting the custom builder function in the config
when creating the reducer’s initial state.  When using all of the
reducers together they make calls to the planTrip action, but there
wasn’t really a way to pass in a custom function that way.  For now it
seems like the only way to do this could be to set something in the
config of the state.

This PR includes the work done in #12.