fix(vehicle-rental): fix place names for eScooter rentals
Previously place names for vehicle (scooter) rentals defaulted to the place name, which often tend
to be UUIDs that have no relevance to the ID found on the scooter/vehicle. This changes the place
name to be $COMPANY $VEHICLE_TYPE and uses the same shared component for bike, car, and
micromobility rental.

fix ibi-group/trimet-mod-otp#213