fix(location-field): log statement to mysteriously fix IE location select onclick
After some amount of debugging I discovered that for some reason on IE, the only way to select a
geocoded result with a click in the location field was to click on the upper part of the MenuItem
(above the text). But by adding a log statement (first I tried to add the statement in the
locationSelected function) anywhere in the app, it seemed to make the entire MenuItem clickable. I
have no idea why this black magic seems to work consistently and had trouble finding info on the
Internet related to this issue.

fix ibi-group/trimet-mod-otp#237