feat(create-react-app): use CRA to build/serve app
This commit reorganizes the project to use Create React App (more specifically the craco tools built
on top of CRA). This required moving the lib dir to a subdir within src, adding/removing some deps,
and some other tweaks. The CSS is not quite working as of this commit.

BREAKING CHANGE: Replace mastarm with Create React App