fix: critival bug fix merging routes in redux
There was a bug in our reducer that caused route info to be overridden when the fetch routes response came back before the fetch route, overriding the extra details we ask for in fetch route. This deep merge will ensure that nothing gets overridden. The $merge utility that we used before was not deep.